+++ packages/core/CHANGELOG.md
@@ -20,6 +20,7 @@ New feature:
 * schema: new option
 
 Fix:
+* session: remove support for 2nd argument
 * connection.wait: fix race condition with quorum
 * events: fix handling next and error events